+++ front/src/Pages/private/CRUDUser/components/ModalAddUser/index.css
@@ -1,0 +1,15 @@
+.title-modal {
+    font-family: Playfair;
+    font-size: 20px;
+    font-weight: 800;
+    letter-spacing: 3;
+    text-align: center;
+}
+
+.modalAllContent {
+    margin-left: 30px;
+}
+
+.containButton {
+    margin-top: 25px;
+}
